add test

+++ Authenticate_Service/Controllers/ProfileController.cs
@@ -5,6 +5,7 @@ using Contract.Service.Message;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Text.RegularExpressions;
 
 namespace AuthenticateService.API.Controllers
 {
@@ -43,10 +44,23 @@ namespace AuthenticateService.API.Controllers
         [HttpDelete]
         public async Task<IActionResult> DeleteUser(string email)
         {
+            // Validate input
+            if (string.IsNullOrEmpty(email))
+            {
+                return new BadRequestObjectResult(Message.MSG11);
+            }
+
+            // Validate email
+            string emailPattern = @"^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$";
+            if (!Regex.IsMatch(email, emailPattern))
+            {
+                return new BadRequestObjectResult(Message.MSG09);
+            }
+
             var user = await _context.Users.FirstOrDefaultAsync(c => c.Email.Equals(email));
             if (user == null)
             {
-                return BadRequest(Message.MSG30);
+                return BadRequest(Message.MSG01);
             }
 
             _context.Users.Remove(user);
